feat: configure js/jsx storybook
