--- packages/superset-ui-demo-old/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@superset-ui/demo",
-  "version": "0.11.0",
-  "description": "Storybook for Superset UI ✨",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "gh-pages:clean": "rm -rf _gh-pages",
-    "gh-pages:build": "build-storybook -o _gh-pages -c storybook-config",
-    "gh-pages:publish": "git-directory-deploy --directory _gh-pages",
-    "gh-pages": "npm run gh-pages:clean && npm run gh-pages:build && npm run gh-pages:publish && npm run gh-pages:clean",
-    "storybook": "start-storybook -p 9001 -c storybook-config"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/apache-superset/superset-ui.git"
-  },
-  "keywords": [
-    "storybook",
-    "superset",
-    "ui",
-    "visualization",
-    "analysis",
-    "data"
-  ],
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/apache-superset/superset-ui/issues"
-  },
-  "homepage": "https://github.com/apache-superset/superset-ui#readme",
-  "dependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/polyfill": "^7.2.5",
-    "@storybook/addon-knobs": "^4.0.2",
-    "@storybook/addon-options": "^4.0.3",
-    "@storybook/react": "^4.0.2",
-    "@superset-ui/chart": "^0.11.0",
-    "@superset-ui/color": "^0.11.0",
-    "@superset-ui/connection": "^0.11.0",
-    "@superset-ui/legacy-plugin-chart-sankey": "^0.10.0",
-    "@superset-ui/legacy-plugin-chart-sunburst": "^0.10.0",
-    "@superset-ui/legacy-plugin-chart-word-cloud": "^0.10.0",
-    "@superset-ui/legacy-preset-chart-big-number": "^0.10.0",
-    "@superset-ui/number-format": "^0.11.0",
-    "@superset-ui/plugin-chart-word-cloud": "^0.10.0",
-    "@superset-ui/time-format": "^0.11.0",
-    "babel-loader": "^8.0.4",
-    "bootstrap": "^4.2.1",
-    "git-directory-deploy": "^1.5.1",
-    "prop-types": "^15.6.2",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "storybook-addon-jsx": "^5.4.0"
-  },
-  "devDependencies": {
-    "@types/storybook__addon-knobs": "^4.0.4"
-  }
-}
