--- packages/superset-ui-demo-old/storybook-config/addons.js
@@ -1,4 +1,0 @@
-// note that the import order here determines the order in the UI!
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-options/register';
-import 'storybook-addon-jsx/register';
