--- packages/superset-ui-demo-old/storybook/stories/superset-ui-chart/index.ts
@@ -1,5 +1,0 @@
-import ChartDataProviderStories from './ChartDataProviderStories';
-
-export default {
-  examples: [...ChartDataProviderStories],
-};
