--- packages/superset-ui-demo-old/storybook/stories/superset-ui-connection/index.js
@@ -1,5 +1,0 @@
-import ConnectionStories from './ConnectionStories';
-
-export default {
-  examples: [...ConnectionStories],
-};
