--- packages/superset-ui-demo-old/types/external.d.ts
@@ -1,4 +1,0 @@
-declare module '@superset-ui/legacy-preset-chart-big-number';
-declare module '@superset-ui/legacy-plugin-chart-sankey';
-declare module '@superset-ui/legacy-plugin-chart-sunburst';
-declare module '@superset-ui/legacy-plugin-chart-word-cloud';
