+++ packages/superset-ui-demo/.storybook/addons.js
@@ -1,3 +1,4 @@
 import '@storybook/addon-actions/register';
 import '@storybook/addon-links/register';
 import '@storybook/addon-knobs/register';
+import 'storybook-addon-jsx/register';
