+++ packages/superset-ui-demo/.storybook/config.js
@@ -1,9 +1,25 @@
-import { configure } from '@storybook/react';
+import { addParameters, configure } from '@storybook/react';
 
-// automatically import all files ending in *.stories.js
-const req = require.context('../stories', true, /\.stories\.js$/);
-function loadStories() {
-  req.keys().forEach(filename => req(filename));
+addParameters({
+  options: {
+    name: '✨ Superset UI',
+    addonPanelInRight: false,
+    enableShortcuts: false,
+    goFullScreen: false,
+    hierarchySeparator: /\|/,
+    selectedAddonPanel: undefined, // The order of addons in the "Addon panel" is the same as you import them in 'addons.js'. The first panel will be opened by default as you run Storybook
+    showAddonPanel: true,
+    showSearchBox: false,
+    showStoriesPanel: true,
+    sidebarAnimations: true,
+    sortStoriesByKind: false,
+    url: '#',
+  },
+});
+
+function loadStorybook() {
+  require('./storybook.css');
+  require('../storybook/stories'); // all of the stories
 }
 
-configure(loadStories, module);
+configure(loadStorybook, module);
