+++ packages/superset-ui-demo/storybook/stories/superset-ui-color/CategoricalStories.jsx
